Collection of post-fitment cc-knob corrections
The indexing pin turns out to be off by pretty much exactly
limiting_pin_y. (Oops.) I also reckon the grip could stand to be taller,
and that there's no real need for the stem to be quite so long.

stem_id really is *outrageously* touchy: 0.205 was on the very edge of
"too small" for the first attempts, but after drying the filament a bit
it's clearly too big.